util/platform: add do_socket
do_socket wraps up the complications of setting O_CLOEXEC on the socket
in a cross-platform way. It also allows us to set SO_REUSEADDR when that
ONEFISH_SO_REUSEADDR_HACK compile option has been configured.

ONEFISH_SO_REUSEADDR_HACK is just a hack to make it easier to run lots
of unit tests in a row. A better solution would be dynamically finding
available ports, but I'll do that later.

Signed-off-by: Colin McCabe <cmccabe@alumni.cmu.edu>