util/packed: add fns for working with packed data
When working with packed data being send over the network, you need to
both convert from network byte order, and handle alignment issues. The
functions in util/packed can do both.

In the future it would be nice to somehow detect at compile time if the
structure field we're working with is guaranteed to be aligned, and use
faster aligned operations in that case. This could be done with some
clever macros, but I don't have time for it right now.

Signed-off-by: Colin McCabe <cmccabe@alumni.cmu.edu>