msgr: replace buggy and slow pipe with ev_async
We must call ev_unloop from the messenger thread. Use an ev_async
watcher to notify the messenger thread that it needs to do this. Also,
avoid using a pipe to send information about transactors to be sent.
Making those syscalls is slower than simply accessing memory. So create
an implicit list in the transactor structure and make its head be in the
messenger.

Signed-off-by: Colin McCabe <cmccabe@alumni.cmu.edu>