msgr: many fixes
* msg.c: calloc_msg now assumes that len includes the length of the base
message structure.  This fits in well with a scenario where users
"subclass" by making struct msg the first member of their structure.

* msg.h: add dump_msg_hdr to convert message headers to human-readable
form.

* reserve 0 as a special transactor ID meaning "none"

* msgr: rename msgr->pending_head to msgr->pending_tr_head

* The transactor ID on the remote system is now different than the local
transactor id.  We have to track both so that we can have bidirectional
communication.  It's not desirable to force them into the same namespace
because then avoiding collisions becomes a chore.

Signed-off-by: Colin McCabe <cmccabe@alumni.cmu.edu>