fast_log: allow setting fast log storage
* process context now checks the environment variable FAST_LOG to see
what fast log message types should be stored.

* moved g_fast_log_mgr out of signal.c.  It never really belonged there.
Instead, it now lives in core/process_ctx.c, which  makes more sense.

* now using weak symbols to keep (most of) the dumper functions out of
libutil.  This is important because otherwise we'll end up pulling in a
lot of datatypes and whatnot into libutil that doesn't really belong
there.  This could have been done by function registration, but using
weak symbols means that the linker does it for me, which I like.
Weak symbols should be available on all the platforms we care about, as
far as I know.  If not, we'll change it later.

Signed-off-by: Colin McCabe <cmccabe@alumni.cmu.edu>