msgr: change callback convention a bit
Messenger completion callbacks no longer take a separate message as an
argument.  Instead, the tr->m field is set to the value that would have
formerly been passed in to the callback.  tr->m was formerly unused
except when a message was queued for sending.  Using it for this purpose
makes sense and does not conflict with anything.

When reporting messenger send failures, do so by setting tr->m to an
error pointer, rather than creating an MMM_NETFAIL message.  Allocating
a message to represent the failure was cumbersome and rather unecessary,
given that we only needed to pass a small error code back.  Allocating
an MMM_NETFAIL message is also an operation which could theoretically
fail, which creates awkward issues of error handling for error handlers
themselves.  It's easier just to pass an error pointer.

Signed-off-by: Colin McCabe <cmccabe@alumni.cmu.edu>