mstor: add mstor_create config param
Add a configuration parameter called 'mstor_create.'  If this is false,
we will not create a new mstor if one does not already exist.  This
definitely makes sense for the fishmdump tool, which normally is meant
to inspect an existing mstor, not create a new one.  If mstor_create is
unset, it defaults to being true.

Signed-off-by: Colin McCabe <cmccabe@alumni.cmu.edu>