mstor: change op-specific calling convention
Pass both the leaf node and the parent of the leaf node to the
operation-specific code in the mstor.  We need both for things like
stat, which has to check for read permissions on the parent directory,
and creat, which can create a new entry that overwrites a previous one.

Signed-off-by: Colin McCabe <cmccabe@alumni.cmu.edu>