client: split disconnect/free, close/free
We need to split off the operation of disconnecting a client from the
operation of freeing the client.  Similarly, we need to split off the
operation of closing a file from the operation of freeing the file data
structure.

This is necessary because of the needs of the Hadoop JNI bindings.  We
need to support a thread-safe close() operation for RedfishClient,
RedfishFileSystem, RedfishInputStream, and RedfishOutputStream.  The
free operations will never be thread-safe.  They invalidate pointers
which other threads may be using.  They also free the mutexes which we
are using for synchronization in the first place.  However, the close()
operations can be made thread-safe.

One other alternative is to make all the JNI methods synchronized, so
that nobody can yank the client or file pointers out from under us.
However, this is undesirable from a performance point of view.
Still another alternative is to rely on the users not to call the close
methods while other threads are using the class.  However, this would
open a hole which malicious users could use to crash the JVM itself.

Signed-off-by: Colin McCabe <cmccabe@alumni.cmu.edu>