client: mess with client shutdown, again
For the "real" Redfish client, the redfish_file structure will contain a
pointer to the redfish_client that created it.  It's possible that the
redfish_file structures created by the client will not be freed until
after the client itself has been freed.

This is an especially big issue for the Hadoop bindings, because the
RedfishDataOutputStream objects do not contain a pointer to the
redfish_client itself, but only to the redfish_file.  We don't want to
be chasing invalid pointers if the RedfishFileSystem object is somehow
garbage collected before a RedfishDataOutputStream object it created.

So we reference count the redfish_client structure.  To reflect this
change, rename redfish_free_client to redfish_release_client.
Update the doxygen comment to reflect the fact that the memory may not
be freed immediately.

NOTE: This is a non-issue for the stub client, because the redfish_file
structures of the stub client don't contain a pointer to the client that
created them.

Signed-off-by: Colin McCabe <cmccabe@alumni.cmu.edu>