fishc: redfish_connect now takes a conf file arg
Rather than taking a list of metadata server ports and hostnames,
redfish_connect now takes a path to a configuration file.

stest: be more verbose if we don't have a mandatory parameter that we
know we need.  Previously, we were only failing after we had already
daemonized, which wasn't ideal.

Signed-off-by: Colin McCabe <cmccabe@alumni.cmu.edu>