msgr: split shutdown into shutdown + free
* Split msgr_shutdown into msgr_shutdown and msgr_free.  This design
choice came out of some issues encountered when building a blocking RPC
framework.  Basically, there was no reasonable way to ever shut down the
threads doing the blocking I/O if the shutdown function invalidated
pointers to the msgr.  It might be possible, although sloppy, to ignore
this issue in daemon code, but the messenger also goes into the client
library.

* Fix a case where pending transactor could fall through the cracks on
shutdown and form a resource leak.

* mtran_deliver_netfail: when reporting connection errors, execute the
messenger callback with conn == NULL.  There probably isn't anything
useful that the API user can do with a connection that's about to be
freed.

* add some comments about locking

* add a unit test of msgr_shutdown delivering ECANCELED.

Signed-off-by: Colin McCabe <cmccabe@alumni.cmu.edu>