msgr: fix a major bug
Previously, we were assuming that a given TCP connection could either be
sending or receiving data at any given time, but never both.  However,
it's completely reasonable for calls to send() to be interleaved with
calls to recv().  Sockets are full duplex, after all.

In particular, there isn't any reason to ever turn off notification of
incoming data on the socket.  We could easily fall into race conditions
if we turned off notification of incoming data right before it was
received on the socket.  (And indeed I observed this kind of problem.)

To fix this, replace the old state-machine based system with a simpler
and more flexible system based on keeping track of the amount of data
that has been sent and the amount of data that has been received.  We
can still only send one message at a time and only receive one message
at a time-- but now the send and receive can be interleaved.

Signed-off-by: Colin McCabe <cmccabe@alumni.cmu.edu>