fuse: fix argument parsing
Parse arguments in the form '-o foo=bar' rather than using getopt as
usual.  Fuse-based daemons are intended to be run by the mount program,
which expects filesystem-specific options to be contained in the string
passed to -o.

This also eliminates worries about our argument parsing colliding with
FUSE's built-in argument parsing.

Signed-off-by: Colin McCabe <cmccabe@alumni.cmu.edu>