fishc: refactor get file status functions
Create redfish_get_file_status, which gets the status of an open file.

Change redfish_stat so that it doesn't contain a path, to make that
easier.  This means that redfish_list_directory must return a different
type, redfish_dir_entry-- a type which does contain a path.

This is good all around, because returning a path for
redfish_get_path_status was just redundant, since the user passed in a
path to identify the file.  Also, since the directory entries now have a
different type, it's more obvious how to correctly use the free_*
functions.

Signed-off-by: Colin McCabe <cmccabe@alumni.cmu.edu>