msgr: rework timeouts
Timeouts used to be handled on a per-TCP-connection basis.  This was
unsatisfactory because we have only one TCP connection open to our MDS
peers, and we want to send messages that have a short timeout along with
messages that have a longer timeout.  (Messenger timeouts are not simply
about delivering packets, but about getting responses to messages.) This
change introduces the concept of a per-transactor timeout.

For performance reasons, we don't want to call clock_gettime every time
we send a message.  So we express timeouts in terms of 'timeo_id,' a
counter inside the messenger that gets incremented every messenger
timeout period (currently hard-coded to a second long.)  We do handle
wraparound in this counter.

msgr->timeo_id reflects the ID of the current messenger timeout period.
When a message is sent on a transactor, we set the timeo_id of that
transactor to some point in the future.  Every second, we examine all of
the currently open TCP connections, and time out the transactors which
are past their expiration timeout period.  We can do this efficiently
because we maintain a per-connection red-black tree of transactors keyed
on the expiration timeout period (timeo_id).

Signed-off-by: Colin McCabe <cmccabe@alumni.cmu.edu>