msgr: fix bug affecting new connections
This fixes a bug in mconn_create.

If a connection was established immediately by the connect(2) system
call, rather than returning EINPROGRESS, we would not have done
ev_io_start, which might have led to messages lingering in the pending
(send) queue indefinitely without being sent.

It makes no sense to check the conn->pending_head queue in mconn_create,
since we just created that queue and it is empty.  Rather, we should
always do ev_io_start if we are creating a new connection with the
specific purpose of sending out a message.

Signed-off-by: Colin McCabe <cmccabe@alumni.cmu.edu>