bsend: remove bsend_cancel
bsend_cancel was both buggy and unecessary.  The big use case for
bsend_cancel was supposed to be shutting down client libraries, but
shutting down the messenger would work just as well.  When the messenger
is shut down, any running bsend contexts would get ESHUTDOWN and no new
RPCs would be able to be made.

Signed-off-by: Colin McCabe <cmccabe@alumni.cmu.edu>