msgr: reference count messages, fix bug
* Fix a bug where we might not properly handle short reads from of the
message header.

* In the messenger, reference count messages.  This allows us to do a
lot of interesting things, like send the same message to multiple
endpoints.  (We don't modify messages that we send inside the
messenger.)

Messages that are being received should have a reference count of 1
until the user gets his grubby hands on them.  At that point the user
can decide what to do with the message-- release it, reuse it, etc.

Signed-off-by: Colin McCabe <cmccabe@alumni.cmu.edu>