mstor: treat user, group as strings, msg reorg
We should treat users and groups as strings.  Otherwise, the client will
have to cache a lookup table that translates between UID/GIDs and
user/group names.  We can still continue to store users as numbers
internally in the mstor itself, though.

Coalsce mmm_nack and mmm_ack into mmm_resp-- a response with an error
code attached (which may be 0 on success.)  This makes the code simpler.

MMM_GET_MDS_STATUS, MMM_STAT_RESP: drop the 'D' in the message name.
These messages don't always have to be used for debugging.  Also include
the primary mds in the status response.

Add MSG_DYNAMIC_CAST, which tries to validate that a message can be
interpreted as some type T.  This includes checking m->ty and m->len.

Signed-off-by: Colin McCabe <cmccabe@alumni.cmu.edu>